:recycle: refactor about ViewGraph+Visitor
