+++ Sources/SwiftTUI/View/Text/Text.swift
@@ -168,3 +168,27 @@ extension Text {
         return self
     }
 }
+
+extension Text: HasIntrinsicContentSize {
+    private func calcTextSize(proposedWidth: PhysicalDistance) -> Size {
+        let contents = content.split(separator: "\n").map { String($0) }
+        let baseHeight = contents.count
+        guard let maxWidthString = contents.max (by: { $0.width < $1.width }) else {
+            return Size(width: proposedWidth, height: baseHeight)
+        }
+        let width = maxWidthString.width
+        if width > proposedWidth {
+            let lineBreakCount = PhysicalDistance(roundf(Float(width) / Float(proposedWidth)))
+            return Size(width: proposedWidth, height: lineBreakCount)
+        }
+        return Size(width: width, height: baseHeight)
+    }
+    func intrinsicContentSize(viewGraph: ViewGraph, visitor: ViewSetRectVisitor) -> Size {
+        if viewGraph.proposedSize.width == 0 {
+            return .zero
+        }
+        let size = calcTextSize(proposedWidth: viewGraph.proposedSize.width)
+        return size
+    }
+}
+
